--- config.json
@@ -1,5 +1,0 @@
-{
-    "apiKey": "xx--your-api-key--abc",
-    "org": "your-org-id",
-    "source": "your-source-id"
-}
